New feature : Slack notification (#100)
Send a notification on a Slack channel when a mark is added or updated by the
logged user. It is also possible to publish a mark and a comment for the current
user with a Publish button on the Movie page.

The Slack notifications are configurable using the Cineapp notification engine.