Use decorator orm.reconstructor
In order to be sure that the attribute guest is initialized with a value in
order to avoid an exception. This is used only when we query a user from the
database and which doesn't have this attribute initialized since it is not
stored in the database.

See: http://docs.sqlalchemy.org/en/latest/orm/constructors.html

Mapping imposes no restrictions or requirements on the constructor (__init__)
method for the class. You are free to require any arguments for the function
that you wish, assign attributes to the instance that are unknown to the ORM,
and generally do anything else you would normally do when writing a constructor
for a Python class.

The SQLAlchemy ORM does not call __init__ when recreating objects from database
rows. The ORM’s process is somewhat akin to the Python standard
library’s pickle module, invoking the low level __new__ method and then
quietly restoring attributes directly on the instance rather than calling
__init__.

Fixes #87