Improve upload_avatar
It is now an edit_profile test which reupdates all notifications
parameters considering the data into database. If we don't do that, the
form is posted with None for all notifications because they are not
correctly initialized and then the notifications are disbaled which is
not good for the next tests.