Dynamic messages and labels displays.
Since we have two modes, we must display differents messages on all
application pages and notifications. The messages.py contains all the
strings to be displayed considering the mode.

In the before_request, function we define in the g object the message
dictionnary to use for messages display.

Relates to: #121