Make general list compatible with tvshows.
Updating all queries in order to handle tvshows and movies. This is
applicable for classical list display, filter search and FTS search.

For the FTS search, it's mandatory to use the inherited object. We can't
do a filtered search on the generic object.

Queries have been also updated for sorting operations.

Code of update_datatable has been migrated into the show blueprint.

Relates to: #121