Convert graph display to hybrid mode (Movie or TvShow)
All queries have been updated in order to filter the displayed data
considering only the active show type. Moreover, graph we can display
are also now filtered since some graphs are not relevant considering the
mode.

For each graph, an option is added in the dictionnary to tell if the
graph can be displayed or not. If False, we return a 404 error.

Relates to: #121