Convert random function to hybrid mode.
random function has to be rewriten: In the previous mode with the movie
table, we were using the table length for the values range for the
random fetch value operation ==> All the values between 1 and length
were possible values.

With the new mode with tvshows and movies in the same table, we can't
use that behaviour anymore: All the values are not usable anymore
because some ids are for movies and other for tvshows. So it can take a
very long time to find an id matching an object with the correct
show_type.

New behaviour is to fetch into a list the available ids and to take an
index randomly in that table. In that way we can be sure that we will
fetch randomly an object/row matching with the current active type.

Function has also been moved into the show blueprint and routes link
have been updated.

Relates to: #121