Add production status information for TV shows.
We display now the production status for the TV shows. Since it's in
english, we use a table in order to do the translation with a status
more understandable for users.

We also add in that table, a style information in order to display a
different color for the displayed button.

Relates to: #121