Change navbar search form generation
Previously, SearchForm was generated and stored in the g object to be
available from any routes and template. It seems it's a very ugly way to
do.

Better way to do is to use Context processors
(https://flask.palletsprojects.com/en/stable/templating/#context-processors)

It allows to generate a dict before create the template.

That commit:
 - Remove the use of g object
 - Generate and return the search form using context processor
 - Update the concerned route in order to use the new form

Fixes #150