Major application refactoring using Flask Factory pattern
In order to be able to test the application with several configurations
files, that refactoring is needed in order to make the application
configurable when it is initalized.

With the previous archiecture, since everything was into the __init__
file, there was only a unique configuration way. Now, it is possible to
pass parameters using the create_app function.

Reference documentation:
https://flask.palletsprojects.com/en/stable/patterns/appfactories/

Major changes :
 - Reorganize __init__.py (create_app function, import new blueprints)
 - Update all python routes and third-party functions
 - Update templates

Ticket #152